Rename graphs.py to visualizations.py
